[tasks] Allow unauthenticated access to public projects
Now we can finally show roadmaps publicly.
